Feature recommendations

+++ UserService/src/main/java/com/streamster/userservice/model/Preferences.java
@@ -9,7 +9,7 @@ public class Preferences {
 
     private Set<String> studyPrograms;
     private Set<String> tags;
-    private Long minLength;
-    private Long maxLength;
+    private Integer minLength;
+    private Integer maxLength;
 
 }
